+++ src/ai-bundle/CHANGELOG.md
@@ -14,6 +14,7 @@ CHANGELOG
  * The `api_version` option for `Azure` store is now `null` by default to allow the usage of a `ScopingHttpClient`
  * The `vector_field` option for `Azure` store is now `vector` by default
  * Add support for `ScopingHttpClient` usage in `AzureSearch` store
+ * Add `similarity_search` option to retriever configuration to automatically register a `SimilaritySearch` tool with customizable `prompt_template`
 
 0.6
 ---